Install in one go
I cleaned up your install command. git is already in the image from upstream, apt-utils aren't really needed in our case since they mostly do stuff that we don't have in docker anyway.

I also moved ssh key generation to runtime, our privates should not be in the image. To ensure that they stay stable we could generate them on the host and add them as single file volumes. That way the keys aren't world-readable and we can still keep them for longer than a single docker run.

I removed stopping of sshd since there is no way it could be running, It's not like we're calling init or whatnot.

Last but not least, I added a call to apt-get clean to ensure that the RUN layer stays as small as possible. We really don't want any package descriptions and caches in the image.